Piles are first-class in state.
State contains top-level piles key. The value is a sorted map indexed by
the x-position of each pile, and each value of this map is another
sorted map indexed by the y-position of each pile at that x-position.
The values of the second (y-indexed) sorted map are the piles, which are
just vectors of cards.