Render selected cards with a selected class.
During rendering, identify selected piles, and for those piles, find the
ids of selected cards and render those cards with a "selected" class to
allow them to be highlighted through styling.

Since this is designed to be used only with cards of a uniform size,
remove the :width and :height keys from card maps, and instead use the
card-width and card-height global constants.