Refactor card selection & rendering.
Add a :selected? fields to cards, since after a selection is applied the
fact that the card has been selected can only be determined by reifying
it in the state.

Change the `selected-cards` fn to `pile-after-selection`, which rather
than returning a sequence of only the selected cards, returns a new
version of the pile it was given with the :selected? field of cards that
are hit by the selection set to true.

Change `render-card` to accept only the card to be rendered as an
argument, and have it look at the :selected? field of the card in order
to determine whether to give it the selected class.