Add within?, for determining if point is in box.
Add the within? function, that takes the min x, max x, min y, and
max y values of a box (i.e. left, right, top, and bottom sides) and the
x and y coordinates of a point, and returns a boolean indicating whether
the point is (inclusively) within the box.

Use within? in pile-selected?.