Make piles in state row-major.
The piles were organized in the state in column-major order, but that
was not a good match. Because cards in a pile stack vertically, column
locations are entirely predictable, but row locations vary. When the
piles were column-major, it was fairly expensive to figure out the
locations of the rows, as it required mapcatting all the piles together
and then finding all the distinct y values, losing all the benefit of
having the piles indexed geometrically in the first place.

With row-major piles, the pile locations can be determined simply by
looking at the keys of the outer (i.e. y pos to row) pile map.