Add API ns with booster and sealed pool routes.
Add the arcane-lab.api ns, which defines a handler that serves
/booster/:set and /pool/:pack-spec routes (where `pack-spec` is
something like 2rtr2gtc2dgm), along with the static resources.

Add a symlink to the compiled js from the resources/js folder, and move
the react.js file to resources/js/static.

Add a ring handler spec to the project.clj so a development server can
be started using `lein ring server`.