Add redo on 'r'.
Replace the too-narrow state history atom with the grander "fate atom",
which contains both the past and the future. If a new drop is made,
though, the future is reset.

Need to refactor rewind! to not be terrible and full of concurrency bugs
(that won't actually happen, because javascript, but still).