Allow space for UI bar at top.
Change the app container div to have relative position, rejigger the
left position to keep it against the left side of the window, and bump
up the top position to give the UI bar space.

Adjust the mouse position signal in the app to compensate for the space
on top.