Add import namespace with card spec fns.
Add the `arcane-lab.import` namespace, which primarily contains
"spec->cards", a function to go from a card spec like "4 Stinkweed Imp"
to a sequence of the appropriate number card maps.

It also contains some functions to recognize when a card spec could not
be parsed and transform such a parsing error into a readable error
message to display to the user.