Add `booster-printings` and `most-recent-printing`
Add two functions to the cards namespace to help with finding particular
card printings.

`booster-printings` is just like `printings`, but it only looks at sets
that were sold in booster packs.

`most-recent-printing` takes a sequence of cards and returns the most
recently printed card (taking advantage of the fact that multiverse IDs
are numeric and strictly monotonically increasing).