Add printings-in-set fn, use in printing-in-set.
Most cards are only printed once in a set, but some (mostly basic lands)
are printed multiple times. Change the old `printing-in-set` function to
`printings-in-set`, by having it return all matching cards from the set
(rather than just the first). Add a new `printing-in-set` that just
grabs the first card from the sequence returned by `printings-in-set`.