Add cost->colortype, use when sorting sealed pools
Battle for Zendikar introduces cards that have colored mana costs but
are colorless (with devoid), which means that we can't just look at the
card's colors in order to figure out which column to put it in for the
initial sealed pool layout -- since you still have to spend colored mana
to cast them, we still want them to be in that color's column.

To that end, introduce a `cost->colortype` function which looks at
a card's mana cost in order to figure out its color. If the card has one
colored mana symol in its cost, it's that color, and if it has multiple,
then it's `:gold`. If it has not color in its cost, then the function
returns nil.

Use the `cost->colortype` function when sorting cards for the sealed
pool layout. If `cost->colortype` doesn't return a color type, then look
at the card's colors; this correctly sorts cards like the pacts or
Evermind, which don't have colors in their cost but have a color
indicator instead.