Add extraneous card predicates.
Some sets have cards that are technically in the set, but don't appear
in booster packs, for whatever reason. Add a mapping between set codes
and predicates that identify the extraneous cards in that set (their
behavior on cards not in that set are undefined), and use it when
processing booster sets to remove these cards from the booster version
of the set (so they won't show up in booster packs).