Print boosters from print-runs.
Add the `print-booster` function, which takes a print run and prints
a new booster pack from it, adding it to the run's :boosters vector. If
the print run is empty, it is replenished before printing. The updated
print run is returned.

Add `print-boosters`, which takes a print run, prints multiple boosters
and adds them all to to the run, returning the updated run.

Change `booster` to create a print run and use `print-booster` to
generate the booster pack it returns.

Change `pool` to make print runs for all the sets in the pool and use
`print-booster` to incrementally print new boosters, with a 1/n chance
(where `n` is the total number of boosters from that set in the pool)
after each booster to switch to a new print run of the same set, and
a 100% chance to switch after three boosters have been printed from the
same run, to reflect how boosters in a pool aren't always taken from the
same print run.