Support double-faced cards.
Add a `:dfc?` field to cards as they're read in, based on whether the
:layout field's value indicates it's a double-faced card (DFC).

When a set is in the self-declared list of DFC sets (which excludes
Origins in order to keep it working), add a :dfcs rarity to the set's
cards map, and remove the DFC cards from the other rarities.

Add a `dfcs-in-boosters?` sets, which is used to see whether the booster
pack spec contains a slot for double faced cards in particular.

For sets in our internal DFC sets list, add a :dfcs sheet to the print
runs, which uses the ISD / DKA ratios for DFC rarities, and add the rest
of the print-run related utilities to get us cards from :double-faced
slots in boosters for those sets.