Support SOI booster packs.
This commit adds support for printing SOI booster packs, but doesn't
update the cards-by-set.json file to include SOI, so the packs can't
actually be printed quite yet.

SOI introduces a new kind of booster slot: the [:double-faced :common]
slot, where there's a 1 in 8 chance of getting a double-faced card, and
a 7 in 8 chance of getting a regular common. In order to support these
"stochastic" slots we need a way to determine which sheet that slot will
pull from when making each booster pack.

To that end, replace the slot->sheet map with a function that does the
same thing as slot->sheet for all the supported slots that map to
a single sheet, and does a sample based on the provided seed to
determine the sheet for stochastic slots, and throws an error if given
a slot it doesn't recognize (whereas before you'd get nil).

Change the `print-booster` function to provide a seed to `slot->sheet`
and use the new sheets definition when selecting all of the sheets from
a print run.