Add `arcane-lab.sets` CLJC ns.
The namespace contains Arcane Lab-specific set metadata, namely about
which sets work and which don't, to control the formats that show up in
the navigator. In the future there might be format information for the
sets defined here that would get merged into the main set data loaded in
`arcane-lab.cards`.