Refactor UI bar.
Rather than some bare links, make a real bar that stays fixed to the top
of the browser, has a distinct background, has a higher z-index than the
card area, and has actual buttons in it.

Reorder the components of the bar (moving the import decklist button
to the end), group like components, rename 'Help' to '?' and 'Import
List' to 'Import Decklist'.