Only update selected cards when change likely.
Rather than finding all the selected cards in the render stage every
time the mouse moves, calculate the selected cards in the action fn
returned by `update-selection-or-drag-destination-action` and put it in
the state, and only re-calculate if the mouse has moved over an
x-coordinate or y-coordinate where a card edge could exist. Essentially,
this puts the selection update behind a bloom filter.

Remove all selection logic from `render-state` and the functions it
calls, since that's now all handled in the state actions and stored in
the state.