Add migrate-state fn, have it also remake piles
Since I've added things to the state that won't be in the states saved
in users' local storage, I've been just adding them to the state when
starting the app, but at this point it makes sense to break out into its
own function.

The function now also re-makes all the piles in the state, in order to
add the height that's now cached in the piles when they're created.