Allow cards to be inserted at any slot in a pile
When the top of the pile of dragged cards is over a targeted placed
pile, then a marker will appear to show where the cards will be inserted
if dropped. If the top of the dragged pile is above or below the
targeted pile, then the cards will be inserted on the top of that pile
(at the end, over the currently-visible card).