Optimize drag/selection dirty trigger checks.
Improve the complexity of checking for a drag or selection update line
between the last two frames' mouse positions from O(n) to O(log_2(n)),
where n is the number of triggering lines. This is achieved by sorting
the trigger lines when caching them in the state, and then doing binary
search index lookups in there for the two frames' coordinates; if the
two lookups yield different indices, then it's time to update.