Fix bug in drag drop zone calc w/empty pile.
An empty pile had an erroneously small drop zone because it had height
0, which it shouldn't when there's others cards in its row. This led
to difficulty putting cards in new piles in the top row and the rares
erroneously falling down to the second row when clicked.

This is the first bug in the drag's drop zone calculation that was
discovered because it didn't match up with the seperately-implemented
drop zone trigger calculation for the drag optimization.