Improve drop highlight behavior, color
Previously the highlight to help you find cards you may have dropped in
a pile would go over *any* individual card, no matter what sort of slot
you dropped it into.

Turns out this not useful when you drop the card into a blank slot or
the top slot in a pile, since those cards are hard to lose, and putting
the highlight over it makes it less pleasing to look at. So, stop doing
that.

Since the highlight only goes over card names, not art or rules text,
and card names are high-contrast, we can crank up the highlight a bit to
a very saturated purple with a slightly higher opacity.