Sort pool by color identity, not color.
This is a better way to sort for most everything besides DFCs, which
sometimes have a different color on the reverse side. Since that doesn't
change what kind of deck you put it in (and so far no DFCs have anything
like off-color activation costs that would lead to a multicolor color
identity from just one side), so those are sorted based on the front
side's colors only.