(maint) Add loading gif
This turned out to be fairly complicated, since I need the application
to recognize when all the images are loaded and remove the spinner,
since there's a few seconds between the application rendering the
initial layout to the DOM and the images all loading, so I can't just
put it in the static HTML and have the application overwrite it when it
loads.

Add the 'imagesloaded' JS library to detect when all the images contained
in an element (& its children) have finished loading.

Use Om's IDidMount in the app's root Om component to tell when there's
actually images in the DOM to point the imagesloaded library at, and
pass it a callback that will set a flag to indicate that the initial
loading is finished so the loading gif should no longer be rendered.

Unfortunately, since Zelkova manages the Om state atom, I can't put the
flag there as it just gets removed the next time a Zelkova event comes
through. Instead use the ugly hack of defining a separate atom in main
just to hold that flag, and passing that to the root render call along
with the om state.