Put DFC backs under green selection borders.
Re-order backs of DFC cards and the main drag pile such that the DFCs
appear to be under the green selection border of the cards in the
dragged pile. When there's at least one non-DFC card above the DFCs in
the drag pile, it's confusing to not have any visual separation between
the top card and the back of a DFC below it, since it appears like the
DFC back corresponds with the top card, which doesn't have a DFC. Having
the green border between them should make it more obvious, and there
will still be no border between the front and back of a DFC when it is
dragged by itself without being selected.