(perf) Switch from using top & left to transform
Rather than usng the 'top' & 'left' CSS properties to position UI
elements, use the 'transform' property with a 'transform($left, $top)'
value instead. This will make things a little faster when dragging
cards, since changing the transform won't trigger a layout operation
like changing top & left does. When the user is creating a selection,
the width & height of the selection div changes, which still triggers
a re-layout.

Adding a transform creates a new stacking context which has consequences
for the z-index, so an explicit z-index needed to be added to the div
that holds the DFC backs when dragging in order to get the backs to stay
above the other cards and the drop target.

Set the loader GIF to be absolutely positioned so the cards will ignore
it when calculating the initial position that the transform translation
applies to; without this, the cards would appear below the loader GIF
instead of on top of it.