Change internal booster representation
Rather than being a flat sequence of the cards in the booster, they're
maps keyed by rarity with sequences of all the cards of that rarity. The
API will flatten them back down by default, but if the `structured`
query param is supplied (and set to anything besides a falsey-ish value)
then the map versions are returned instead.