Change most frontend logic namespaces to be cljc
I guess I did a good job of structuring things when I refactored out the
frontend into a bunch of different namespaces, because most of those
namespaces were pure clojure and were moved over to be cljc namespaces
with no changes at all.

The only thing that needed to be changed was, ironically enough, in the
constants namespace: replace the dynamic calculation of timestamp digits
with a hardcoded constant (that will be accurate for the rest of my
lifetime so whatever).