Add cljc UI ns for creating initial UI state
The fruit of moving all those namespaces from cljs to cljc in the last
commit is being able to move the creation of the initial UI state given
a sealed pool over to cljc as well (with the end goal of pre-rendering
the initial state on the server so the user will see cards sooner).

Add a new `arcane-lab.ui` namespace that contains the functions related
to this initial UI state calculation, which are all refactored out of
the `sealed-pool-piles` and `deck-piles` functions that were in
main.cljs. It also has the functions that transform cards from the API
representation to the UI one by adding fields that the UI needs.

Take the opportunity to do a bit of refactoring around how the UI
calculates the initial state by having two separate paths for starting
the app if in constructed deck mode or sealed pool mode. Previously,
both modes used the same callback function on the AJAX request that got the
cards from the API, and the callback had some hacky logic to determine
which mode it was in. Now the two different modes use different callback
functions that don't have to resort to such hacks but instead simply
create the initial piles as appropriate for their mode and then pass
that off to the new `start-app-from-piles!` function.

Since the UI namespace's new `pool->piles` function that replaces the
old `sealed-pool-piles` expects to be given the pool's cards keyed by
the sheet they were printed on (rather than a flat collection of the
cards in the pool), add the `structured` query param to the AJAX request
that gets a sealed pool.