Pre-render initial pool UI state on server
Now I get to really reap the benefits of having the same language on
both the front and back ends because pre-rendering the UI on the server
is simple now that all the logic part is in host-agnostic cljc.

In order to indicate to the user that the app isn't initially
interactible, change the loading indicator from the egg gif that's
covered by the cards to a popover with a neat crawling infinity loader
*SVG* (fancy right?).

Somewhat mixed up in this is changing the "client" nomenclature
convention in the UI ns to "ui"--client is a little too generic of
a word--and changing a bug in the conversion that made all the cards
show up in the upper left slot (the origin slot).