Only placeholder cards in server-rendered pool.
Don't put the actual card images in the pre-rendered pool UI that the
server sends to the client, because this seems to cause Chrome to not
show *any* card images at all until all the images have been loaded,
which defeats most of the purpose of pre-rendering the pool. Instead,
set all the card images to be the card back, and let the clojurescript
set the real card images once it all starts up.