Fix redo button bug
In the transition to rendering the page on the server, I messed up the
redo button on the UI bar and put the button outside the link, which
made it do nothing. I really should have noticed, too, because visually
it was off: the redo button was closer to the new pool button than the
undo button.