Fix saved state / visible state mismatch bug
Previously, the saved state was only updated when a new state was added
to the history (meaning the user made a move). This behavior is
undesirable when the user hits undo and then navigates away or reloads,
since the saved state is not updated by the undo so when they come back
they'll have the state that they wanted to undo, with no undo history.

Fix this by updating the saved state whenever the user hits undo or
redo.