(maint) Simplify selection/single-card drag logic
In `action/start-drag`, we were looping through all the piles with
selected cards, and checking the intersection between the selected cards
in that pile and the mouse down location, and also checking for a card
under the mouse down location with `piles/card-under`. We can greatly
simplify things by replacing the selected piles intersection logic with
a simple check to see if the card returned by `piles/card-under` is
marked as selected or not. This also ends up being more efficient,
especially when the selection includes cards from multiple piles.

Can't believe I didn't think of this sooner; I got sucked into the
complicated intersection checking logic, which is kinda fun to write.