Add (unused) first draft of Cypress logic port
The new arcane-lab.state-machine namespace defines a UI state machine
with Cypress, and its transition functions are a bunch of new, much
simpler functions added to arcane-lab.action alongside the existing
Zelkova signal graph actions.

I did refactor the drag-target function so it would better fit the
cypress action functions, which lift the decision of whether to 'undo'
the drag and put the card back where it came from into the UI state
machine graph, rather than sticking a marker in the app state and
looking for it in the drag-target function. This does have the
side-effect of removing your ability to click on a card and then drop it
back where it came from if you didn't move it, which would normally be
a bug but we're about to rip all the UI logic out and replace it with
the new stuff so who cares.

Next step is to rip the Zelkova logic out and switch over to using the
Cypress state logic, see what blows up and find the bugs.