(maint) Break out drag-target rendering
Apart from the generally good principle of decomposing a larger function
into two smaller ones, a small functional change made during the cypress
switchover is to not render the drag target until the mouse has moved
during the drag. This change allows for that, and having the drag target
be a separate function makes it much easier to skip rendering it if
necessary.