Finish Cypress port and switch over to it.
Add rewind and fast-forward actions to the Cypress state machine and
hook them up to be triggered from idle on 'u' and 'r' key presses and the
:arcane-lab/undo and :arcane-lab/redo custom events. The custom event
part requires upgrading to a new version of Cypress.

In main, set up supplementary core.async channels to emit the custom
events to the Cypress state machine when the undo and redo button DOM
elements get a mouse down event.

Fix some small issues in the first draft of the Cypress actions. They
were almost entirely bug-free!