Add M19 as new default format
This one was non-trivial because of M19's unusual land sheet, which has
common 2-color tap-lands in addition to basics on it, in 40/60
proportion. This is not the first set to do something unusual with the
land sheet, so the framework is already in place for handling this;
I only had to extend it to M19.

I wasn't able to totally confirm this, but based on PVDDR's article and
the fact that there are 10 extra commons in this set as compared to
a normal set (111 vs 101) and that I can't think of a good way to handle
10 additional commons on the commons sheet (20 is easier, but 10 is
tricky), it seems that the tap-lands only show up on the lands sheet and
not the commons sheet. To handle that, make the commons sheet
construction customizable per-set and use this to filter out the
tap-lands when making an M19 commons sheet.