Fix broken navigator bar
Using the raw set meant that the conditional to check whether to augment
a set with its `:sealed-format`, which hinges on whether the set had
boosters, would have to use the raw `:boosterV3` key, not the `:booster`
key we prefer to use. Shadow the 'mtg-set' var with the renamed value in
order to use `:booster` for the check.

Doing this fixes the navigation bar because without knowing the sets'
sealed formats, the navigation bar can neither recognize which format
it's in nor send you to new ones.