Proxy card images from Scryfall, w/in-mem cache
The image responses (including the bodies) are cached in memory. They're
~150 KB per card, so a whole set's worth of images is only ~40 MB, and
most of the requests will be for the most recent set. Should make it LRU
with a size cap to eliminate the DoS vector.

As Scryfall is the only image source left, remove the whole multi-source
scaffolding.

This shouldn't be used as-is because it'll request images from Scryfall
faster then they ask you to (they will remind & then block IPs that send
requests more frequently than one every 100 ms). Need to figure out some
way to rate limit the Scryfall requests.