Throttle scryfall requests to one per 100 ms
Use the new `arcane-lab.schedule` namespace to coordinate the scheduling
of requests amongst the threads. A schedule is just an atom that has
a vector with maps that record a scheduled request :time and the ID of
the requesting :thread (the maps are ordered in the vector by :time).
Each thread uses the schedule to find the next open slot (i.e. the next
millisecond where no requests have been made in the previous 100 ms),
marks that slot in the schedule, and then sleeps until the slot time.

Tested it out by logging the time that each card image request was made
when loading a sealed pool and manually inspecting those timestamps to
verify that at least 100 ms passed between each one.