Expand `str->long`'s support, add docstring
Previously, only strings that were entirely integers yielded a non-nil
result. Now, if the string starts with an integer, that integer is
returned and the rest of the string is ignored.

Since this might not always be the best choice for the caller, it now
has a docstring that calls this out.