Handle weird collector numbers found on some cards
The MTGJSON data is now a lot more comprehensive, and it contains cards
with weird collector numbers that don't even start with an integer. This
was causing exceptions because `parse-collector-number` assumed that all
collector numbers at least started with an integer.

Change `parse-collector-number` to use `str->long` to see if the
collector number string starts with an integer; if not, it's returned
unmodified.